refactor: change percentile method to calculate several
Change the percentile method to receive an iterator with the percentiles
to calculate and calculate all of them rather than having a method that
only receive one percentile to calculate at a time.

Because of this change, now the percentile sub-command has a default of
calculating the 50th and 99th percentiles.